backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1

# AI/LangChain
langchain==0.1.9
langchain-google-genai==0.0.11
langchain-core==0.1.27

# Kafka
confluent-kafka[avro]==2.3.0
fastavro>=1.9.0

# Google APIs
google-ads==24.1.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
gspread==6.0.2

# Database (PostgreSQL)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# WebSocket
python-socketio==5.11.1

# Utils
httpx==0.27.0
aiofiles==23.2.1
cryptography==42.0.2
